requirements.txt
# Core ETL
beautifulsoup4==4.14.2
pydantic==2.12.3
pydantic_settings==2.11.0
python-dotenv==1.2.1
Requests==2.32.5
selenium==4.38.0
undetected_chromedriver==3.5.5
webdriver_manager==4.0.2

# Database
supabase==2.24.0
SQLAlchemy==2.0.35

# Analytics & Visualization
pandas==2.2.0
plotly==5.18.0
streamlit==1.31.0
yfinance==0.2.48

# Data Quality
great-expectations==0.18.8

# API & Async Tasks
fastapi==0.115.0
uvicorn[standard]==0.32.0
celery==5.4.0
redis==5.2.0
flower==2.0.1